[FIX] demo_mn_console: Fix compiler warning (2)
The merge of the develop-2.2 branch in the develop-2.3
reverted the change introduced by the commit
c7cd3e170de912e78465a1fe1f1de0b9d6007670 with the changes
from b5c8cc8a7ca92702f9cf5adbafd1fc06d71f8a79.

The patch fixes the merge issue.

+++ apps/demo_mn_console/src/app.c
@@ -269,9 +269,9 @@ static tOplkError initProcessImage(void)
     tOplkError      ret = kErrorOk;
 
     printf("Initializing process image...\n");
-    printf("Size of process image: Input = %ld Output = %ld\n", sizeof(PI_IN), sizeof(PI_OUT));
+    printf("Size of process image: Input = %lu Output = %lu\n", (ULONG)sizeof(PI_IN), (ULONG)sizeof(PI_OUT));
     eventlog_printMessage(kEventlogLevelInfo, kEventlogCategoryGeneric,
-                        "Allocating process image: Input:%d Output:%d", sizeof(PI_IN), sizeof(PI_OUT));
+                        "Allocating process image: Input:%lu Output:%lu", (ULONG)sizeof(PI_IN), (ULONG)sizeof(PI_OUT));
 
     ret = oplk_allocProcessImage(sizeof(PI_IN), sizeof(PI_OUT));
     if (ret != kErrorOk)
